// src/modules/users/interface/grpc/dto/user-response.dto.ts
export class UserResponseDto {
  id: string;
  idNumber: string;
  idType: string;
  name: string;
  nickname: string;
  preferredUsername: string;
  picture: string;
  email: string;
  emailVerified: string;
  gender: string;
  birthdate: string;
  phoneNumber: string;
  phoneNumberVerified: string;
  address: string;
  placeOfBirth: string;
  nationalities: string;
  postalCode: string;
  city: string;
  createdAt: string;
  updatedAt: string;
  status: string;
}